Collapse trend wizard/runner modules into trend.py
trend_wizard.py and trend_runner.py are now private functions within
trend.py. The injectable-callback test seam (ask_from/ask_to) is preserved.
No behaviour changes.
